+++ tck/communication-tck/communication-tck-document/src/main/java/jakarta/nosql/communication/tck/document/DeleteQueryBuilderTest.java
@@ -15,123 +15,69 @@
  */
 package jakarta.nosql.communication.tck.document;
 
+
 import jakarta.nosql.Condition;
-import jakarta.nosql.Sort;
-import jakarta.nosql.SortType;
 import jakarta.nosql.TypeReference;
 import jakarta.nosql.document.Document;
 import jakarta.nosql.document.DocumentCollectionManager;
 import jakarta.nosql.document.DocumentCollectionManagerAsync;
 import jakarta.nosql.document.DocumentCondition;
-import jakarta.nosql.document.DocumentEntity;
-import jakarta.nosql.document.DocumentQuery;
+import jakarta.nosql.document.DocumentDeleteQuery;
 import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Mockito;
+import org.mockito.ArgumentMatchers;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.function.Consumer;
-import java.util.stream.Stream;
 
 import static jakarta.nosql.document.DocumentCondition.eq;
-import static jakarta.nosql.document.DocumentQuery.select;
+import static jakarta.nosql.document.DocumentDeleteQuery.delete;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 
-
-public class DefaultSelectQueryBuilderTest {
-
+public class DeleteQueryBuilderTest {
 
     @Test
     public void shouldReturnErrorWhenHasNullElementInSelect() {
-        assertThrows(NullPointerException.class, () -> select("document", "document'", null));
+        Assertions.assertThrows(NullPointerException.class,() -> delete("document", "document", null));
     }
 
     @Test
-    public void shouldSelect() {
+    public void shouldDelete() {
         String documentCollection = "documentCollection";
-        DocumentQuery query = select().from(documentCollection).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).build();
         assertTrue(query.getDocuments().isEmpty());
         assertFalse(query.getCondition().isPresent());
         assertEquals(documentCollection, query.getDocumentCollection());
     }
 
     @Test
-    public void shouldSelectDocument() {
+    public void shouldDeleteDocuments() {
         String documentCollection = "documentCollection";
-        DocumentQuery query = select("document", "document2").from(documentCollection).build();
+        DocumentDeleteQuery query = delete("document", "document2").from(documentCollection).build();
         assertThat(query.getDocuments(), containsInAnyOrder("document", "document2"));
         assertFalse(query.getCondition().isPresent());
         assertEquals(documentCollection, query.getDocumentCollection());
     }
 
+
     @Test
     public void shouldReturnErrorWhenFromIsNull() {
-        assertThrows(NullPointerException.class, () -> select().from(null));
-    }
-
-
-    @Test
-    public void shouldSelectOrderAsc() {
-        String documentCollection = "documentCollection";
-        DocumentQuery query = select().from(documentCollection).orderBy("name").asc().build();
-        assertTrue(query.getDocuments().isEmpty());
-        assertFalse(query.getCondition().isPresent());
-        assertEquals(documentCollection, query.getDocumentCollection());
-        assertThat(query.getSorts(), Matchers.contains(Sort.of("name", SortType.ASC)));
-    }
-
-    @Test
-    public void shouldSelectOrderDesc() {
-        String documentCollection = "documentCollection";
-        DocumentQuery query = select().from(documentCollection).orderBy("name").desc().build();
-        assertTrue(query.getDocuments().isEmpty());
-        assertFalse(query.getCondition().isPresent());
-        assertEquals(documentCollection, query.getDocumentCollection());
-        assertThat(query.getSorts(), contains(Sort.of("name", SortType.DESC)));
-    }
-
-
-    @Test
-    public void shouldReturnErrorSelectWhenOrderIsNull() {
-        assertThrows(NullPointerException.class,() -> {
-            String documentCollection = "documentCollection";
-            select().from(documentCollection).orderBy(null);
-        });
-    }
-
-    @Test
-    public void shouldSelectLimit() {
-        String documentCollection = "documentCollection";
-        DocumentQuery query = select().from(documentCollection).limit(10).build();
-        assertTrue(query.getDocuments().isEmpty());
-        assertFalse(query.getCondition().isPresent());
-        assertEquals(documentCollection, query.getDocumentCollection());
-        assertEquals(10L, query.getLimit());
-    }
-
-    @Test
-    public void shouldSelectSkip() {
-        String documentCollection = "documentCollection";
-        DocumentQuery query = select().from(documentCollection).skip(10).build();
-        assertTrue(query.getDocuments().isEmpty());
-        assertFalse(query.getCondition().isPresent());
-        assertEquals(documentCollection, query.getDocumentCollection());
-        assertEquals(10L, query.getSkip());
+        Assertions.assertThrows(NullPointerException.class,() -> delete().from(null));
     }
 
     @Test
     public void shouldSelectWhereNameEq() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentQuery query = select().from(documentCollection).where("name").eq(name).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").eq(name).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -148,7 +94,7 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameLike() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentQuery query = select().from(documentCollection).where("name").like(name).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").like(name).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -164,7 +110,7 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameGt() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentQuery query = select().from(documentCollection).where("name").gt(value).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").gt(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -180,7 +126,7 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameGte() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentQuery query = select().from(documentCollection).where("name").gte(value).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").gte(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -196,7 +142,7 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameLt() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentQuery query = select().from(documentCollection).where("name").lt(value).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").lt(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -212,7 +158,7 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameLte() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentQuery query = select().from(documentCollection).where("name").lte(value).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").lte(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -229,7 +175,7 @@ public class DefaultSelectQueryBuilderTest {
         String documentCollection = "documentCollection";
         Number valueA = 10;
         Number valueB = 20;
-        DocumentQuery query = select().from(documentCollection).where("name").between(valueA, valueB).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").between(valueA, valueB).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -238,15 +184,14 @@ public class DefaultSelectQueryBuilderTest {
         assertEquals(documentCollection, query.getDocumentCollection());
         assertEquals(Condition.BETWEEN, condition.getCondition());
         assertEquals("name", document.getName());
-        assertThat(document.get(new TypeReference<List<Number>>() {
-        }), Matchers.contains(10, 20));
+        assertThat(document.get(new TypeReference<List<Number>>() {}), Matchers.contains(10, 20));
     }
 
     @Test
     public void shouldSelectWhereNameNot() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentQuery query = select().from(documentCollection).where("name").not().eq(name).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").not().eq(name).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document column = condition.getDocument();
@@ -264,7 +209,7 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameAnd() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentQuery query = select().from(documentCollection).where("name").eq(name).and("age")
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").eq(name).and("age")
                 .gt(10).build();
         DocumentCondition condition = query.getCondition().get();
 
@@ -280,7 +225,8 @@ public class DefaultSelectQueryBuilderTest {
     public void shouldSelectWhereNameOr() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentQuery query = select().from(documentCollection).where("name").eq(name).or("age").gt(10).build();
+        DocumentDeleteQuery query = delete().from(documentCollection).where("name").eq(name)
+                .or("age").gt(10).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -292,10 +238,11 @@ public class DefaultSelectQueryBuilderTest {
     }
 
 
+
     @Test
-    public void shouldSelectNegate() {
+    public void shouldDeleteNegate() {
         String columnFamily = "columnFamily";
-        DocumentQuery query = select().from(columnFamily).where("city").not().eq("Assis")
+        DocumentDeleteQuery query = delete().from(columnFamily).where("city").not().eq("Assis")
                 .and("name").not().eq("Lucas").build();
 
         DocumentCondition condition = query.getCondition().orElseThrow(RuntimeException::new);
@@ -308,56 +255,49 @@ public class DefaultSelectQueryBuilderTest {
         assertThat(conditions, containsInAnyOrder(eq(Document.of("city", "Assis")).negate(),
                 eq(Document.of("name", "Lucas")).negate()));
 
-    }
 
-
-    @Test
-    public void shouldExecuteManager() {
-        DocumentCollectionManager manager = Mockito.mock(DocumentCollectionManager.class);
-        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
-        String collection = "collection";
-        Stream<DocumentEntity> entities = select().from(collection).getResult(manager);
-        Mockito.verify(manager).select(queryCaptor.capture());
-        checkQuery(queryCaptor, collection);
     }
 
     @Test
-    public void shouldExecuteSingleResultManager() {
-        DocumentCollectionManager manager = Mockito.mock(DocumentCollectionManager.class);
-        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
+    public void shouldExecuteDelete() {
         String collection = "collection";
-        Optional<DocumentEntity> entities = select().from(collection).getSingleResult(manager);
-        Mockito.verify(manager).singleResult(queryCaptor.capture());
-        checkQuery(queryCaptor, collection);
-    }
+        DocumentCollectionManager manager = mock(DocumentCollectionManager.class);
+        ArgumentCaptor<DocumentDeleteQuery> queryCaptor = ArgumentCaptor.forClass(DocumentDeleteQuery.class);
+        delete().from(collection).delete(manager);
+        verify(manager).delete(queryCaptor.capture());
 
-    @Test
-    public void shouldExecuteManagerAsync() {
-        DocumentCollectionManagerAsync manager = Mockito.mock(DocumentCollectionManagerAsync.class);
-        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
-        String collection = "collection";
-        Consumer<Stream<DocumentEntity>> callback = System.out::println;
-        select().from(collection).getResult(manager, callback);
-        Mockito.verify(manager).select(queryCaptor.capture(), Mockito.eq(callback));
-        checkQuery(queryCaptor, collection);
-    }
-
-    @Test
-    public void shouldExecuteSingleResultManagerAsync() {
-        DocumentCollectionManagerAsync manager = Mockito.mock(DocumentCollectionManagerAsync.class);
-        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
-        String collection = "collection";
-        Consumer<Optional<DocumentEntity>> callback = System.out::println;
-        select().from(collection).getSingleResult(manager, callback);
-        Mockito.verify(manager).singleResult(queryCaptor.capture(), Mockito.eq(callback));
-        checkQuery(queryCaptor, collection);
-    }
-
-    private void checkQuery(ArgumentCaptor<DocumentQuery> queryCaptor, String collection) {
-        DocumentQuery query = queryCaptor.getValue();
+        DocumentDeleteQuery query = queryCaptor.getValue();
         assertTrue(query.getDocuments().isEmpty());
         assertFalse(query.getCondition().isPresent());
         assertEquals(collection, query.getDocumentCollection());
     }
 
+    @Test
+    public void shouldExecuteAsyncDelete() {
+        String collection = "collection";
+        DocumentCollectionManagerAsync manager = mock(DocumentCollectionManagerAsync.class);
+        ArgumentCaptor<DocumentDeleteQuery> queryCaptor = ArgumentCaptor.forClass(DocumentDeleteQuery.class);
+        delete().from(collection).delete(manager);
+        verify(manager).delete(queryCaptor.capture());
+
+        DocumentDeleteQuery query = queryCaptor.getValue();
+        assertTrue(query.getDocuments().isEmpty());
+        assertFalse(query.getCondition().isPresent());
+        assertEquals(collection, query.getDocumentCollection());
+    }
+
+    @Test
+    public void shouldExecuteAsync2Delete() {
+        String collection = "collection";
+        DocumentCollectionManagerAsync manager = mock(DocumentCollectionManagerAsync.class);
+        ArgumentCaptor<DocumentDeleteQuery> queryCaptor = ArgumentCaptor.forClass(DocumentDeleteQuery.class);
+        Consumer<Void> callback = (v) ->{};
+        delete().from(collection).delete(manager, callback);
+        verify(manager).delete(queryCaptor.capture(), ArgumentMatchers.eq(callback));
+
+        DocumentDeleteQuery query = queryCaptor.getValue();
+        assertTrue(query.getDocuments().isEmpty());
+        assertFalse(query.getCondition().isPresent());
+        assertEquals(collection, query.getDocumentCollection());
+    }
 }
