+++ tck/communication-tck/communication-tck-document/src/main/java/jakarta/nosql/communication/tck/document/SelectQueryBuilderTest.java
@@ -15,69 +15,123 @@
  */
 package jakarta.nosql.communication.tck.document;
 
-
 import jakarta.nosql.Condition;
+import jakarta.nosql.Sort;
+import jakarta.nosql.SortType;
 import jakarta.nosql.TypeReference;
 import jakarta.nosql.document.Document;
 import jakarta.nosql.document.DocumentCollectionManager;
 import jakarta.nosql.document.DocumentCollectionManagerAsync;
 import jakarta.nosql.document.DocumentCondition;
-import jakarta.nosql.document.DocumentDeleteQuery;
+import jakarta.nosql.document.DocumentEntity;
+import jakarta.nosql.document.DocumentQuery;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
-import org.mockito.ArgumentMatchers;
+import org.mockito.Mockito;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Consumer;
+import java.util.stream.Stream;
 
 import static jakarta.nosql.document.DocumentCondition.eq;
-import static jakarta.nosql.document.DocumentDeleteQuery.delete;
+import static jakarta.nosql.document.DocumentQuery.select;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 
-public class DefaultDeleteQueryBuilderTest {
+
+public class SelectQueryBuilderTest {
+
 
     @Test
     public void shouldReturnErrorWhenHasNullElementInSelect() {
-        Assertions.assertThrows(NullPointerException.class,() -> delete("document", "document", null));
+        assertThrows(NullPointerException.class, () -> select("document", "document'", null));
     }
 
     @Test
-    public void shouldDelete() {
+    public void shouldSelect() {
         String documentCollection = "documentCollection";
-        DocumentDeleteQuery query = delete().from(documentCollection).build();
+        DocumentQuery query = select().from(documentCollection).build();
         assertTrue(query.getDocuments().isEmpty());
         assertFalse(query.getCondition().isPresent());
         assertEquals(documentCollection, query.getDocumentCollection());
     }
 
     @Test
-    public void shouldDeleteDocuments() {
+    public void shouldSelectDocument() {
         String documentCollection = "documentCollection";
-        DocumentDeleteQuery query = delete("document", "document2").from(documentCollection).build();
+        DocumentQuery query = select("document", "document2").from(documentCollection).build();
         assertThat(query.getDocuments(), containsInAnyOrder("document", "document2"));
         assertFalse(query.getCondition().isPresent());
         assertEquals(documentCollection, query.getDocumentCollection());
     }
 
-
     @Test
     public void shouldReturnErrorWhenFromIsNull() {
-        Assertions.assertThrows(NullPointerException.class,() -> delete().from(null));
+        assertThrows(NullPointerException.class, () -> select().from(null));
+    }
+
+
+    @Test
+    public void shouldSelectOrderAsc() {
+        String documentCollection = "documentCollection";
+        DocumentQuery query = select().from(documentCollection).orderBy("name").asc().build();
+        assertTrue(query.getDocuments().isEmpty());
+        assertFalse(query.getCondition().isPresent());
+        assertEquals(documentCollection, query.getDocumentCollection());
+        assertThat(query.getSorts(), Matchers.contains(Sort.of("name", SortType.ASC)));
+    }
+
+    @Test
+    public void shouldSelectOrderDesc() {
+        String documentCollection = "documentCollection";
+        DocumentQuery query = select().from(documentCollection).orderBy("name").desc().build();
+        assertTrue(query.getDocuments().isEmpty());
+        assertFalse(query.getCondition().isPresent());
+        assertEquals(documentCollection, query.getDocumentCollection());
+        assertThat(query.getSorts(), contains(Sort.of("name", SortType.DESC)));
+    }
+
+
+    @Test
+    public void shouldReturnErrorSelectWhenOrderIsNull() {
+        assertThrows(NullPointerException.class,() -> {
+            String documentCollection = "documentCollection";
+            select().from(documentCollection).orderBy(null);
+        });
+    }
+
+    @Test
+    public void shouldSelectLimit() {
+        String documentCollection = "documentCollection";
+        DocumentQuery query = select().from(documentCollection).limit(10).build();
+        assertTrue(query.getDocuments().isEmpty());
+        assertFalse(query.getCondition().isPresent());
+        assertEquals(documentCollection, query.getDocumentCollection());
+        assertEquals(10L, query.getLimit());
+    }
+
+    @Test
+    public void shouldSelectSkip() {
+        String documentCollection = "documentCollection";
+        DocumentQuery query = select().from(documentCollection).skip(10).build();
+        assertTrue(query.getDocuments().isEmpty());
+        assertFalse(query.getCondition().isPresent());
+        assertEquals(documentCollection, query.getDocumentCollection());
+        assertEquals(10L, query.getSkip());
     }
 
     @Test
     public void shouldSelectWhereNameEq() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").eq(name).build();
+        DocumentQuery query = select().from(documentCollection).where("name").eq(name).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -94,7 +148,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameLike() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").like(name).build();
+        DocumentQuery query = select().from(documentCollection).where("name").like(name).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -110,7 +164,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameGt() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").gt(value).build();
+        DocumentQuery query = select().from(documentCollection).where("name").gt(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -126,7 +180,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameGte() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").gte(value).build();
+        DocumentQuery query = select().from(documentCollection).where("name").gte(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -142,7 +196,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameLt() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").lt(value).build();
+        DocumentQuery query = select().from(documentCollection).where("name").lt(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -158,7 +212,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameLte() {
         String documentCollection = "documentCollection";
         Number value = 10;
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").lte(value).build();
+        DocumentQuery query = select().from(documentCollection).where("name").lte(value).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -175,7 +229,7 @@ public class DefaultDeleteQueryBuilderTest {
         String documentCollection = "documentCollection";
         Number valueA = 10;
         Number valueB = 20;
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").between(valueA, valueB).build();
+        DocumentQuery query = select().from(documentCollection).where("name").between(valueA, valueB).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -184,14 +238,15 @@ public class DefaultDeleteQueryBuilderTest {
         assertEquals(documentCollection, query.getDocumentCollection());
         assertEquals(Condition.BETWEEN, condition.getCondition());
         assertEquals("name", document.getName());
-        assertThat(document.get(new TypeReference<List<Number>>() {}), Matchers.contains(10, 20));
+        assertThat(document.get(new TypeReference<List<Number>>() {
+        }), Matchers.contains(10, 20));
     }
 
     @Test
     public void shouldSelectWhereNameNot() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").not().eq(name).build();
+        DocumentQuery query = select().from(documentCollection).where("name").not().eq(name).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document column = condition.getDocument();
@@ -209,7 +264,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameAnd() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").eq(name).and("age")
+        DocumentQuery query = select().from(documentCollection).where("name").eq(name).and("age")
                 .gt(10).build();
         DocumentCondition condition = query.getCondition().get();
 
@@ -225,8 +280,7 @@ public class DefaultDeleteQueryBuilderTest {
     public void shouldSelectWhereNameOr() {
         String documentCollection = "documentCollection";
         String name = "Ada Lovelace";
-        DocumentDeleteQuery query = delete().from(documentCollection).where("name").eq(name)
-                .or("age").gt(10).build();
+        DocumentQuery query = select().from(documentCollection).where("name").eq(name).or("age").gt(10).build();
         DocumentCondition condition = query.getCondition().get();
 
         Document document = condition.getDocument();
@@ -238,11 +292,10 @@ public class DefaultDeleteQueryBuilderTest {
     }
 
 
-
     @Test
-    public void shouldDeleteNegate() {
+    public void shouldSelectNegate() {
         String columnFamily = "columnFamily";
-        DocumentDeleteQuery query = delete().from(columnFamily).where("city").not().eq("Assis")
+        DocumentQuery query = select().from(columnFamily).where("city").not().eq("Assis")
                 .and("name").not().eq("Lucas").build();
 
         DocumentCondition condition = query.getCondition().orElseThrow(RuntimeException::new);
@@ -255,49 +308,56 @@ public class DefaultDeleteQueryBuilderTest {
         assertThat(conditions, containsInAnyOrder(eq(Document.of("city", "Assis")).negate(),
                 eq(Document.of("name", "Lucas")).negate()));
 
+    }
 
+
+    @Test
+    public void shouldExecuteManager() {
+        DocumentCollectionManager manager = Mockito.mock(DocumentCollectionManager.class);
+        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
+        String collection = "collection";
+        Stream<DocumentEntity> entities = select().from(collection).getResult(manager);
+        Mockito.verify(manager).select(queryCaptor.capture());
+        checkQuery(queryCaptor, collection);
     }
 
     @Test
-    public void shouldExecuteDelete() {
+    public void shouldExecuteSingleResultManager() {
+        DocumentCollectionManager manager = Mockito.mock(DocumentCollectionManager.class);
+        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
         String collection = "collection";
-        DocumentCollectionManager manager = mock(DocumentCollectionManager.class);
-        ArgumentCaptor<DocumentDeleteQuery> queryCaptor = ArgumentCaptor.forClass(DocumentDeleteQuery.class);
-        delete().from(collection).delete(manager);
-        verify(manager).delete(queryCaptor.capture());
+        Optional<DocumentEntity> entities = select().from(collection).getSingleResult(manager);
+        Mockito.verify(manager).singleResult(queryCaptor.capture());
+        checkQuery(queryCaptor, collection);
+    }
 
-        DocumentDeleteQuery query = queryCaptor.getValue();
+    @Test
+    public void shouldExecuteManagerAsync() {
+        DocumentCollectionManagerAsync manager = Mockito.mock(DocumentCollectionManagerAsync.class);
+        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
+        String collection = "collection";
+        Consumer<Stream<DocumentEntity>> callback = System.out::println;
+        select().from(collection).getResult(manager, callback);
+        Mockito.verify(manager).select(queryCaptor.capture(), Mockito.eq(callback));
+        checkQuery(queryCaptor, collection);
+    }
+
+    @Test
+    public void shouldExecuteSingleResultManagerAsync() {
+        DocumentCollectionManagerAsync manager = Mockito.mock(DocumentCollectionManagerAsync.class);
+        ArgumentCaptor<DocumentQuery> queryCaptor = ArgumentCaptor.forClass(DocumentQuery.class);
+        String collection = "collection";
+        Consumer<Optional<DocumentEntity>> callback = System.out::println;
+        select().from(collection).getSingleResult(manager, callback);
+        Mockito.verify(manager).singleResult(queryCaptor.capture(), Mockito.eq(callback));
+        checkQuery(queryCaptor, collection);
+    }
+
+    private void checkQuery(ArgumentCaptor<DocumentQuery> queryCaptor, String collection) {
+        DocumentQuery query = queryCaptor.getValue();
         assertTrue(query.getDocuments().isEmpty());
         assertFalse(query.getCondition().isPresent());
         assertEquals(collection, query.getDocumentCollection());
     }
 
-    @Test
-    public void shouldExecuteAsyncDelete() {
-        String collection = "collection";
-        DocumentCollectionManagerAsync manager = mock(DocumentCollectionManagerAsync.class);
-        ArgumentCaptor<DocumentDeleteQuery> queryCaptor = ArgumentCaptor.forClass(DocumentDeleteQuery.class);
-        delete().from(collection).delete(manager);
-        verify(manager).delete(queryCaptor.capture());
-
-        DocumentDeleteQuery query = queryCaptor.getValue();
-        assertTrue(query.getDocuments().isEmpty());
-        assertFalse(query.getCondition().isPresent());
-        assertEquals(collection, query.getDocumentCollection());
-    }
-
-    @Test
-    public void shouldExecuteAsync2Delete() {
-        String collection = "collection";
-        DocumentCollectionManagerAsync manager = mock(DocumentCollectionManagerAsync.class);
-        ArgumentCaptor<DocumentDeleteQuery> queryCaptor = ArgumentCaptor.forClass(DocumentDeleteQuery.class);
-        Consumer<Void> callback = (v) ->{};
-        delete().from(collection).delete(manager, callback);
-        verify(manager).delete(queryCaptor.capture(), ArgumentMatchers.eq(callback));
-
-        DocumentDeleteQuery query = queryCaptor.getValue();
-        assertTrue(query.getDocuments().isEmpty());
-        assertFalse(query.getCondition().isPresent());
-        assertEquals(collection, query.getDocumentCollection());
-    }
 }
